Add isomorphic-unfetch package
Adding it in anticipation of fetching data in `getInitialProps`.

We could have simply added node-fetch package to add support for
fetch API on the server-side.

But we did not. Why?

This is because we want to support fetch on both server & client sides.
isomorphic-fetch auto switches between node-fetch (server-side) and
unfetch (client-side) packages depending on environment.

But why must we support fetch on both sides?

This is because once SSR pages are rendered in browser, the client-side
React code takes over. A user may use browser back/forward buttons
to navigate, and so the client-side React code may need to remount
the component. Hence, the need to fetch data again (but on client-side).